backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
azure-cosmos>=4.7.0
azure-identity>=1.17.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
python-multipart>=0.0.9
openai>=1.40.0
semantic-kernel>=1.0.0
azure-ai-contentsafety>=1.0.0
